finish part4

+++ part4/bloglist/controller/blog.js
@@ -1,8 +1,19 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+//     return authorization.substring(7)
+//   }
+//   return null
+// }
+
 blogRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
   } catch (exception) {
     next(exception)
@@ -10,26 +21,47 @@ blogRouter.get('/', async (request, response, next) => {
 })
 
 blogRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body)
-  if(!blog.title && !blog.url) {
-    response.status(400).end()
-  } else  {
+
+  const token = response.token.resultToken
+
+  try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = new Blog(request.body)
+    blog.user = user._id
+    if(!blog.title && !blog.url) {
+      response.status(400).end()
+    }
     if(!blog.likes) {
       blog.likes =0
     }
-    try {
-      const blogSaved = await blog.save()
-      response.status(201).json(blogSaved)
-    } catch (exception) {
-      next(exception)
-    }
+    const blogSaved = await blog.save()
+    user.blogs = user.blogs.concat(blogSaved._id)
+    await user.save()
+    response.status(201).json(blogSaved.toJSON())
+  }catch (exception) {
+    next(exception)
   }
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const token = response.token.resultToken
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if(blog.user.toString() === decodedToken.id) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'blog can be deleted only by the user who added the blog' })
+    }
+
   } catch (exception) {
     next(exception)
   }
